Remove ui8 with updated XUSG, since it is not standard C++.

+++ DXRVoxelizer/Content/Voxelizer.h
@@ -21,7 +21,7 @@ public:
 	void Render(const XUSG::RayTracing::CommandList* pCommandList, uint32_t frameIndex,
 		const XUSG::Descriptor& rtv, const XUSG::Descriptor& dsv);
 
-	static const uint32_t FrameCount = 3;
+	static const uint8_t FrameCount = 3;
 
 protected:
 	enum PipelineLayoutIndex : uint8_t
