+++ DXRVoxelizer/DXRVoxelizer.h
@@ -44,7 +44,7 @@ public:
 	virtual void ParseCommandLineArgs(wchar_t* argv[], int argc);
 
 private:
-	static const uint32_t FrameCount = Voxelizer::FrameCount;
+	static const uint8_t FrameCount = Voxelizer::FrameCount;
 
 	// Pipeline objects.
 	XUSG::Viewport			m_viewport;
